net, setup, netpod: Extract bridge binding discovery funcs to a file
Main functionality that handles the discovery of the bridge binding is
moved to a separate file.

The separation is aimed to assist in the code structure, making clear
which par of the code is related to bridge binding.
